+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased
 
  - Replaced the `Invert` selection indicator with `Rectangle`
  - Renamed `SelectValue::name` to `marker`
+ - Renamed `MenuItem` trait to `MenuListItem`
 
 ## Removed